Move createRawSeries to the shared visualizations lib
