+++ frontend/src/embedding-sdk-bundle/lib/sdk-question/update-question.ts
@@ -3,10 +3,10 @@ import _ from "underscore";
 import type { SdkQuestionState } from "embedding-sdk-bundle/types/question";
 import { computeQuestionPivotTable } from "metabase/query_builder/actions/core/pivot-table";
 import { getAdHocQuestionWithVizSettings } from "metabase/query_builder/actions/core/utils";
-import { createRawSeries } from "metabase/query_builder/utils";
 import { loadMetadataForCard } from "metabase/questions/actions";
 import type { Dispatch, GetState } from "metabase/redux/store";
 import { getMetadata } from "metabase/selectors/metadata";
+import { createRawSeries } from "metabase/visualizations/lib/series";
 import * as Lib from "metabase-lib";
 import Question from "metabase-lib/v1/Question";
 import type { ParameterValuesMap } from "metabase-types/api";
